requirements.txt
plotly-express==0.4.1
SQLAlchemy==1.4.14
streamlit==0.81.1